requirements.txt
aldjemy==0.7.0
amqp==2.1.4
appdirs==1.4.3
asn1crypto==0.22.0
attrs==16.3.0
Automat==0.5.0
berserker-resolver==2.0.1
billiard==3.5.0.2
boto3==1.4.4
botocore==1.5.44
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
constantly==15.1.0
coreapi==2.3.0
coreschema==0.0.4
cryptography==1.8.1
cssselect==1.0.1
Django==1.11
django-cors-headers==2.0.2
django-filter==1.0.2
djangorestframework==3.6.2
dnspython==1.15.0
docutils==0.13.1
editdistance==0.3.1
elasticsearch==5.3.0
enum34==1.1.6
Faker==0.7.11
futures==3.1.1
Geohash==1.0
idna==2.5
incremental==16.10.1
ipaddress==1.0.18
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.2
kombu==4.0.2
lxml==3.7.3
MarkupSafe==1.0
mock==1.3.0
nassl==0.16.3
netaddr==0.7.19
netifaces==0.10.5
olefile==0.44
packaging==16.8
parsel==1.1.0
Pillow==4.1.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
queuelib==1.4.2
redis==2.10.5
requests==2.13.0
s3transfer==0.1.10
Scrapy==1.3.3
selenium==3.4.1
service-identity==16.0.0
six==1.10.0
SQLAlchemy==1.1.9
SSLyze==1.1.0
stripe==1.53.0
Twisted==17.1.0
typing==3.6.1
uritemplate==3.0.0
urllib3==1.21
vine==1.1.3
w3lib==1.17.0
XlsxWriter==0.9.6
zope.interface==4.4.0
